algo: completed vo algo

+++ twinhook/th_config.h
@@ -1,12 +1,12 @@
 #pragma once
 
 namespace th_param {
-	const int WINDOW_WIDTH = 640;
-	const int WINDOW_HEIGHT = 480;
-	const int GAME_WIDTH = 384;
-	const int GAME_HEIGHT = 448;
-	const int GAME_X_OFFSET = 32;
-	const int GAME_Y_OFFSET = 16;
+	const float WINDOW_WIDTH = 640;
+	const float WINDOW_HEIGHT = 480;
+	const float GAME_WIDTH = 384;
+	const float GAME_HEIGHT = 448;
+	const float GAME_X_OFFSET = 32;
+	const float GAME_Y_OFFSET = 16;
 
 	const vec2 PLAYER_SIZE(5, 5);
 }
